+++ backend/server/community.go
@@ -10,7 +10,7 @@ func GetCommunityList() ([]*models.Community, error) {
 	return mysql.GetCommunityList()
 }
 
-func GetCommunityDetailByID(id uint) (*models.CommunityDetail, error) {
+func GetCommunityDetailByID(id uint64) (*models.CommunityDetail, error) {
 	// 查找数据，查找到所有的community 并返回
 	return mysql.GetCommunityDetailByID(id)
 }
